+++ src/components/sect1CardComponent.jsx
